refactor(api): claude_sessions 백그라운드 DB 동기화를 sync 모듈로 추출
_sync_sessions_to_db · scan_and_sync_claude_snapshots · _sync_cache · _sync_lock
을 sync.py로 이동. 라우트가 없는 유일한 모듈이며 비-라우트 공개 진입점을
제공한다 (external_usage.py:141 이 직접 호출).

R4b — test_claude_session_sync.py 의 patch.object 대상을 _legacy → sync 로
이동했다. 이 스텝을 빼면 패치는 성공하고 호출은 원본을 타서 테스트가
거짓 통과한다. 추출과 같은 커밋에 둔다.

core(list_sessions)가 _sync_sessions_to_db 를 fire-and-forget 으로 호출하는
것이 이 패키지의 유일한 모듈 간 의존 엣지다 — 공용 심볼이 하나뿐이라
_shared.py 를 만들지 않고 직접 import 한다.

원문 구간 diff 로 바이트 동일 확인. 라우트 테이블 25개 동일.
