+++ tests/backend/test_claude_session_sync.py
@@ -7,7 +7,7 @@
 그대로 두어 재노출 계약까지 함께 검증한다.
 
 분할 진행에 따라 이 대상은 이동한다: `api/claude_sessions.py`(원본) →
-`_legacy`(패키지 승격) → `sync`(도메인 추출). 대상을 패키지로 되돌리지 마라.
+`_legacy`(패키지 승격) → `sync`(도메인 추출, 현재). 대상을 패키지로 되돌리지 마라.
 """
 
 from __future__ import annotations
@@ -26,7 +26,7 @@ async def test_scan_and_sync_claude_snapshots_discovers_and_upserts() -> None:
     prior visit to the Claude Sessions page (freshness follow-up).
     """
     from api import claude_sessions
-    from api.claude_sessions import _legacy as impl
+    from api.claude_sessions import sync as impl
 
     fake_sessions = [MagicMock(), MagicMock(), MagicMock()]
     monitor = MagicMock()
@@ -46,7 +46,7 @@ async def test_scan_and_sync_claude_snapshots_discovers_and_upserts() -> None:
 async def test_scan_and_sync_claude_snapshots_empty() -> None:
     """No discovered sessions → sync still runs with [] and count is 0."""
     from api import claude_sessions
-    from api.claude_sessions import _legacy as impl
+    from api.claude_sessions import sync as impl
 
     monitor = MagicMock()
     monitor.discover_sessions = MagicMock(return_value=[])
